app/requirements.txt
# Main library to connect with Telegram API
telethon==1.37.0

# Driver to connect Python with PostgreSQL database
psycopg2-binary==2.9.10

# To read the secret variables from the .env file
python-dotenv==1.1.0
